requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
certifi==2026.2.25
cfgv==3.5.0
distlib==0.4.0
distro==1.9.0
filelock==3.25.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.18
idna==3.11
iniconfig==2.3.0
nodeenv==1.10.0
packaging==26.0
platformdirs==4.9.4
playwright==1.58.0
pluggy==1.6.0
pre-commit==4.5.1
pydantic==2.12.5
pydantic-core==2.41.5
pyee==13.0.1
pygments==2.19.2
pytest==9.0.2
python-discovery==1.2.0
pyyaml==6.0.3
ruff==0.15.7
sniffio==1.3.1
stagehand==3.6.0
typing-extensions==4.15.0
typing-inspection==0.4.2
virtualenv==21.2.0
